moved IEmailReadOperations to InterfacesLib
