+++ Ginger/GingerCoreCommon/InterfacesLib/IEmailReadOperations.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using GingerCore.GeneralLib;
 
-namespace GingerCore.GeneralLib
+namespace Amdocs.Ginger.Common.InterfacesLib
 {
     public interface IEmailReadOperations
     {
